Convert the redis support into another orchestration provider.
